+++ test/internal/app/models/user.rb
@@ -1,3 +1,4 @@
 class User < ActiveRecord::Base
-  devise :database_authenticatable
+  devise :database_authenticatable, :registerable,
+         :recoverable, :trailable
 end