+++ test/internal/db/schema.rb
@@ -22,5 +22,7 @@ ActiveRecord::Schema.define do
   create_table :users do |t|
     t.string :email
     t.string :encrypted_password
+    t.string :reset_password_token
+    t.datetime :reset_password_sent_at
   end
 end