+++ test/test_helper.rb
@@ -2,6 +2,7 @@ require "bundler/setup"
 Bundler.require(:development)
 require "minitest/autorun"
 require "minitest/pride"
+require "devise"
 
 Devise.setup do |config|
   require "devise/orm/active_record"
@@ -9,13 +10,17 @@ Devise.setup do |config|
   config.warden do |manager|
     manager.failure_app = ->(env) { [401, {"Content-Type" => "text/html"}, "Unauthorized"] }
   end
+
+  config.mailer_sender = "test@example.com"
 end
 
 Combustion.path = "test/internal"
-Combustion.initialize! :active_record, :action_controller do
+Combustion.initialize! :active_record, :action_controller, :action_mailer do
   if ActiveRecord::VERSION::MAJOR < 6 && config.active_record.sqlite3.respond_to?(:represent_boolean_as_integer)
     config.active_record.sqlite3.represent_boolean_as_integer = true
   end
 end
+
+ActionMailer::Base.delivery_method = :test
 
 AuthTrail.geocode = false